stream-processing-movies-leans-data/app/kafka_producer/requirements.txt
confluent_kafka==2.3.0
python-dotenv==1.0.0
requests==2.31.0